services/lightrag/requirements.txt
lightrag-hku==0.0.0b0
neo4j==5.23.0
python-dotenv==1.0.1
structlog==24.4.0
sentence-transformers==3.0.0
pydantic==2.9.2
pyyaml==6.0.2
rapidfuzz==3.10.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1
respx==0.22.0
